Create Name the Triangle - Python
Asks the user for the length of each of the three sides of a triangle, from shortest to longest. Then tell the user if the triangle is a right triangle, an isosceles triangle, an equilateral triangle, a valid triangle but not of a special type, or not a valid triangle at all.