Create Joe likes to be comfortable - Python
A program using a while loop that starts at 70 degrees and randomly alters the temperature at each iteration of the loop by plus or minus two degrees. Print out the temperature and what Joe is doing at the beginning and after each step. If Joe goes inside, end the loop and the program. Use a break statement to get out of the loop when Joe goes inside.