Create A Cash Register - Python
The cash register simulation should use a loop in which the program repeatedly asks the user which of the products they want to buy, and how many. If the product requested isn’t in the inventory, refuse to sell it and ask again. If the number requested is more than the inventory, refuse to sell it and ask again. After each purchase of some quantity of an item, the program should print the price of the product/quantity combination and a running subtotal, rounded off to cents.