Add files via upload
The game is a variant of the card game “Blackjack”. Two players alternately roll dice, and keep track of their total across turns. They are each trying to reach a sum that lies in a specified target, between a fixed low value and high value. If a player reaches a score in the target range, they immediately win. If they exceed the high value,
they immediately lose. The players can choose the number of dice to roll on each turn, between 1 and a fixed maximum. The learning algorithm is as follows: the machine plays a series of games with itself, playing both sides. Initially, both players play randomly; as the series progresses, they hopefully play better and better.